requirements.txt
Flask==3.0.0
python-docx==1.1.0
reportlab==4.0.7
num2words==0.5.12